feat: Bank & Backpack Redesign — spec + impl (milestones 1 + 2a-2g)
PR #5. SYS-12 delivered end-to-end. 409 unit + 11 integration tests green.

## Round 6: 4 Copilot findings

- SlotGrid callbacks now re-resolve the live ItemStack via
  _inventory.GetSlot(index) at click/focus time instead of capturing the
  snapshot from render time. Inventory replaces ItemStack records on
  update (they're immutable records), so captured values went stale
  after any Refresh — the fix ensures actions operate on current state.
- SlotGrid empty slots are non-interactive: FocusMode=None + Disabled
  + "disabled" stylebox. FocusFirstSlot prefers occupied slots (by
  non-empty Text) and falls back to any focusable slot only if all are
  empty. Matches PauseMenu's inventory-grid keyboard-nav behavior.
- .claude/scheduled_tasks.lock (Claude Code session-local runtime
  state — contains sessionId, pid, timestamp) is now gitignored and
  untracked. Was accidentally committed in round 5.
- (CHANGELOG/dev-journal text discrepancy about "bank-first" vs
  "backpack-first" is resolved by the squash — this commit's text is
  authoritative and matches the implementation in Bank.PurchaseExpansion.)

## Round 5: NumberFormat.Abbrev truncates instead of rounding

Rounding lifted 1_950 to "2.0K" and 99_950 to "100.0K" — both ugly.
Now uses Math.Floor(scaled * 10) / 10 in the <100 range so values
stay within the 1.0-99.9 band as the spec implies. Added 6 theory
test cases locking the truncate behavior in.

## Round 4: 7 Copilot findings

DeathScreen keyboard focus, idol path no longer shows paid Save*
options, Strings.Death constants used in button labels, SlotGrid
init-order fix, SlotGrid XML summary matches impl, Bank.PurchaseExpansion
draws backpack-first, GuildWindow.ShowBuyDialog unused local removed.

## New workflow: notify Claude on Copilot review

.github/workflows/notify-claude-on-copilot-review.yml fires a Claude
Code routine on copilot-pull-request-reviewer[bot] reviews. Requires
CLAUDE_ROUTINE_FIRE_URL + CLAUDE_ROUTINE_TOKEN secrets. Fires once
merged to main.

## Round 3: 4 doc/consistency fixes

docs/inventory/{bank,backpack}.md Current State reflects impl,
guild-window.md MVP deviations fully enumerated, DeathPenalty.PayBuyout
notes auto-backpack-first default.

## Round 2: 10 real code + doc fixes

NumberFormat long.MinValue, Inventory.SetLocked idempotent, DeathScreen
Quit-to-main-menu, BackpackWindow Drop confirmation, GuildWindow
Transfer gold tab rebuild, DeathTests predicate lookups, town roster
+ flows/town.md + systems/death.md consistency, session numbering.

## Round 1: spec + impl

NPC restructure: Guild Maid merges Shopkeeper + Banker; Village Chief
from Guild Master; Old Master Blacksmith; Old Master Wizard. PC =
"{Class} Guildmaster".

Model rewrite (Item/Inventory/Bank/DeathPenalty): long gold + counts,
one-type-per-slot invariant, Locked flag, Transfer/Drop/ToggleLock,
two gold pockets, 50xN linear expansion, new buyout formulas.

SlotGrid + NumberFormat + GuildWindow (Store/Bank/Transfer).
BackpackWindow refactored onto SlotGrid with Use/Lock/Drop. DeathScreen
rewritten for 5-option sacrifice dialog.

Guild Maid placed in town (Banker sprite placeholder until ART-02).

+++ scripts/ui/ShopWindow.cs
@@ -267,7 +267,7 @@ public partial class ShopWindow : GameWindow
         _descStats.Text = stats.ToString();
 
         // Update Buy/Sell button text with price + disable when unaffordable
-        int gold = GameState.Instance.PlayerInventory.Gold;
+        long gold = GameState.Instance.PlayerInventory.Gold;
         if (_isBuyMode)
         {
             _buyButton.Text = Strings.Shop.Buy(item.BuyPrice);
